+++ packages/typespec-python/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release
 
+## 0.63.8
+
+- Bump @typespec/http-client-python to 0.37.3
+
+## 0.63.7
+
+- Bump @typespec/http-client-python to 0.37.2
+
 ## 0.63.6
 
 ### Bug Fixes